fix edge case

+++ wp/gametime/_config.py
@@ -1,4 +1,9 @@
-DF_DTYPES = {"steam_id": str, "game_id": str, "game_time": int, "game_time_diff": float}
+DF_DTYPES = {
+    "steam_id": str,
+    "game_id": str,
+    "game_time": float,
+    "game_time_diff": float,
+}
 
 _STEAM_API_KEY: str = "847931CAE62FC345F305B1F974DC8759"
 _STEAM_BEJEWELED_APP_ID: int = 78000
